List auction types based on 'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'

List auction types based on 'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'

Usage

auction_types

Format

An object of class tbl_df (inherits from tbl, data.frame) with 8 rows and 3 columns.

See also

Examples

head(entsoeapi::auction_types)
#> # A tibble: 6 × 3
#>   code  title          description                                                                       
#>   <chr> <chr>          <chr>                                                                             
#> 1 A01   Implicit       The auction is an implicit auction.                                               
#> 2 A02   Explicit       The auction is an explicit auction.                                               
#> 3 A03   Rule Based     The auction is a rule based auction.                                              
#> 4 A04   Mixed          The auction is partially implicit and partially explicit.                         
#> 5 A05   Explicit/split The auction concerns two explicit auctions on a split border.                     
#> 6 A06   Shadow auction An explicit auction carried out in the case of the failure of an implicit auction.