Get Weekly Average Filling Rate of Water Reservoirs and Hydro Storage Plants (16.1.D)

Aggregated weekly average filling rate of all water reservoir and hydro storage plants (MWh) per area, including the same week value of the previous year.

Usage

gen_storage_mean_filling_rate(
  eic = NULL,
  period_start = ymd(Sys.Date() - days(x = 7L), tz = "CET"),
  period_end = ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)

Arguments

eic

Energy Identification Code of the control area, bidding zone or country

period_start

POSIXct or YYYY-MM-DD HH:MM:SS format Maximum 380 days range limit applies

period_end

POSIXct or YYYY-MM-DD HH:MM:SS format Maximum 380 days range limit applies

tidy_output

Defaults to TRUE. If TRUE, then flatten nested tables.

security_token

Security token for ENTSO-E transparency platform

Value

A tibble::tibble() with the queried data.

See also

Examples

df <- entsoeapi::gen_storage_mean_filling_rate(
  eic          = "10YFR-RTE------C",
  period_start = lubridate::ymd(x = "2020-02-01", tz = "CET"),
  period_end   = lubridate::ymd(x = "2021-02-15", tz = "CET"),
  tidy_output  = TRUE
)
#> 
#> ── API call ────────────────────────────────────────────────────────────────────────────────────────────────────────────
#> → https://web-api.tp.entsoe.eu/api?documentType=A72&processType=A16&in_Domain=10YFR-RTE------C&periodStart=202001312300&periodEnd=202102142300&securityToken=<...>
#> <- HTTP/2 200 
#> <- date: Tue, 31 Mar 2026 06:58:03 GMT
#> <- content-type: text/xml
#> <- content-disposition: inline; filename="Aggregate Filling Rate of Water Reservoirs_202001312300-202102142300.xml"
#> <- x-content-type-options: nosniff
#> <- x-xss-protection: 0
#> <- vary: accept-encoding
#> <- content-encoding: gzip
#> <- strict-transport-security: max-age=15724800; includeSubDomains
#> <- 
#> ✔ response has arrived
#> ✔ Additional type names have been added!
#> ✔ Additional eic names have been added!
#> ✔ Additional definitions have been added!

dplyr::glimpse(df)
#> Rows: 55
#> Columns: 21
#> $ ts_in_bidding_zone_domain_mrid  <chr> "10YFR-RTE------C", "10YFR-RTE------C", "10YFR-RTE------C", "10YFR-RTE------C"…
#> $ ts_in_bidding_zone_domain_name  <chr> "France", "France", "France", "France", "France", "France", "France", "France"…
#> $ type                            <chr> "A72", "A72", "A72", "A72", "A72", "A72", "A72", "A72", "A72", "A72", "A72", "…
#> $ type_def                        <chr> "Reservoir filling information", "Reservoir filling information", "Reservoir f…
#> $ process_type                    <chr> "A16", "A16", "A16", "A16", "A16", "A16", "A16", "A16", "A16", "A16", "A16", "…
#> $ process_type_def                <chr> "Realised", "Realised", "Realised", "Realised", "Realised", "Realised", "Reali…
#> $ ts_object_aggregation           <chr> "A01", "A01", "A01", "A01", "A01", "A01", "A01", "A01", "A01", "A01", "A01", "…
#> $ ts_object_aggregation_def       <chr> "Area", "Area", "Area", "Area", "Area", "Area", "Area", "Area", "Area", "Area"…
#> $ ts_business_type                <chr> "A01", "A01", "A01", "A01", "A01", "A01", "A01", "A01", "A01", "A01", "A01", "…
#> $ ts_business_type_def            <chr> "Production", "Production", "Production", "Production", "Production", "Product…
#> $ created_date_time               <dttm> 2026-03-31 06:58:03, 2026-03-31 06:58:03, 2026-03-31 06:58:03, 2026-03-31 06:…
#> $ revision_number                 <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ time_period_time_interval_start <dttm> 2020-01-26 23:00:00, 2020-01-26 23:00:00, 2020-01-26 23:00:00, 2020-01-26 23:…
#> $ time_period_time_interval_end   <dttm> 2021-02-14 23:00:00, 2021-02-14 23:00:00, 2021-02-14 23:00:00, 2021-02-14 23:…
#> $ ts_resolution                   <chr> "P7D", "P7D", "P7D", "P7D", "P7D", "P7D", "P7D", "P7D", "P7D", "P7D", "P7D", "…
#> $ ts_time_interval_start          <dttm> 2020-01-26 23:00:00, 2020-01-26 23:00:00, 2020-01-26 23:00:00, 2020-01-26 23:…
#> $ ts_time_interval_end            <dttm> 2021-02-14 23:00:00, 2021-02-14 23:00:00, 2021-02-14 23:00:00, 2021-02-14 23:…
#> $ ts_mrid                         <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ ts_point_dt_start               <dttm> 2020-01-26 23:00:00, 2020-02-02 23:00:00, 2020-02-09 23:00:00, 2020-02-16 23:…
#> $ ts_point_quantity               <dbl> 2194236, 2214084, 2174132, 2116585, 2008603, 1864482, 1816423, 1779274, 176397…
#> $ ts_quantity_measure_unit_name   <chr> "MWH", "MWH", "MWH", "MWH", "MWH", "MWH", "MWH", "MWH", "MWH", "MWH", "MWH", "…