Get Generation Forecasts for Wind & Solar (14.1.D)

A respective forecast of wind and solar power net generation (MW) per area and each market time unit of the following/current day. The elements of the result list are representing each related forecast time range.

Usage

gen_wind_solar_forecasts(
  eic = NULL,
  period_start = lubridate::ymd(Sys.Date(), tz = "CET") - lubridate::days(x = 1L),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)

Arguments

eic

Energy Identification Code of the control area, bidding zone or country

period_start

POSIXct or YYYY-MM-DD HH:MM:SS format One year range limit applies

period_end

POSIXct or YYYY-MM-DD HH:MM:SS format One year range limit applies

tidy_output

Defaults to TRUE. If TRUE, then flatten nested tables.

security_token

Security token for ENTSO-E transparency platform

Examples

df_list <- entsoeapi::gen_wind_solar_forecasts(
  eic          = "10YFR-RTE------C",
  period_start = lubridate::ymd(x = "2020-02-01", tz = "CET"),
  period_end   = lubridate::ymd(x = "2020-03-01", tz = "CET"),
  tidy_output  = TRUE
)
#> 
#> ── API call ────────────────────────────────────────────────────────────────────────────────────────────────────────────
#> → https://web-api.tp.entsoe.eu/api?documentType=A69&processType=A01&in_Domain=10YFR-RTE------C&periodStart=202001312300&periodEnd=202002292300&securityToken=<...>
#> <- HTTP/2 200 
#> <- date: Mon, 09 Mar 2026 12:42:14 GMT
#> <- content-type: text/xml
#> <- content-disposition: inline; filename="Generation Forecasts Wind Solar_202001312300-202002292300.xml"
#> <- x-content-type-options: nosniff
#> <- x-xss-protection: 0
#> <- vary: accept-encoding
#> <- content-encoding: gzip
#> <- strict-transport-security: max-age=15724800; includeSubDomains
#> <- 
#> ✔ response has arrived
#> 
#> ── API call ────────────────────────────────────────────────────────────────────────────────────────────────────────────
#> → https://web-api.tp.entsoe.eu/api?documentType=A69&processType=A40&in_Domain=10YFR-RTE------C&periodStart=202001312300&periodEnd=202002292300&securityToken=<...>
#> <- HTTP/2 200 
#> <- date: Mon, 09 Mar 2026 12:42:15 GMT
#> <- content-type: text/xml
#> <- content-disposition: inline; filename="Generation Forecasts Wind Solar_202001312300-202002292300.xml"
#> <- x-content-type-options: nosniff
#> <- x-xss-protection: 0
#> <- vary: accept-encoding
#> <- content-encoding: gzip
#> <- strict-transport-security: max-age=15724800; includeSubDomains
#> <- 
#> ✔ response has arrived
#> 
#> ── API call ────────────────────────────────────────────────────────────────────────────────────────────────────────────
#> → https://web-api.tp.entsoe.eu/api?documentType=A69&processType=A18&in_Domain=10YFR-RTE------C&periodStart=202001312300&periodEnd=202002292300&securityToken=<...>
#> <- HTTP/2 200 
#> <- date: Mon, 09 Mar 2026 12:42:16 GMT
#> <- content-type: text/xml
#> <- content-disposition: inline; filename="Generation Forecasts Wind Solar_202001312300-202002292300.xml"
#> <- x-content-type-options: nosniff
#> <- x-xss-protection: 0
#> <- vary: accept-encoding
#> <- content-encoding: gzip
#> <- strict-transport-security: max-age=15724800; includeSubDomains
#> <- 
#> ✔ response has arrived

dplyr::glimpse(df_list$`Day-ahead`)
#> Rows: 1,392
#> Columns: 23
#> $ ts_in_bidding_zone_domain_mrid  <chr> "10YFR-RTE------C", "10YFR-RTE------C", "10YFR-RTE------C", "10YFR-RTE------C"…
#> $ ts_in_bidding_zone_domain_name  <chr> "France", "France", "France", "France", "France", "France", "France", "France"…
#> $ type                            <chr> "A69", "A69", "A69", "A69", "A69", "A69", "A69", "A69", "A69", "A69", "A69", "…
#> $ type_def                        <chr> "Wind and solar forecast", "Wind and solar forecast", "Wind and solar forecast…
#> $ process_type                    <chr> "A01", "A01", "A01", "A01", "A01", "A01", "A01", "A01", "A01", "A01", "A01", "…
#> $ process_type_def                <chr> "Day ahead", "Day ahead", "Day ahead", "Day ahead", "Day ahead", "Day ahead", …
#> $ ts_object_aggregation           <chr> "A08", "A08", "A08", "A08", "A08", "A08", "A08", "A08", "A08", "A08", "A08", "…
#> $ ts_object_aggregation_def       <chr> "Resource type", "Resource type", "Resource type", "Resource type", "Resource …
#> $ ts_business_type                <chr> "A94", "A94", "A94", "A94", "A94", "A94", "A94", "A94", "A94", "A94", "A94", "…
#> $ ts_business_type_def            <chr> "Solar generation", "Solar generation", "Solar generation", "Solar generation"…
#> $ ts_mkt_psr_type                 <chr> "B16", "B16", "B16", "B16", "B16", "B16", "B16", "B16", "B16", "B16", "B16", "…
#> $ ts_mkt_psr_type_def             <chr> "Solar unspecified", "Solar unspecified", "Solar unspecified", "Solar unspecif…
#> $ created_date_time               <dttm> 2026-03-09 12:42:14, 2026-03-09 12:42:14, 2026-03-09 12:42:14, 2026-03-09 12:…
#> $ revision_number                 <dbl> 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, …
#> $ time_period_time_interval_start <dttm> 2020-01-31 23:00:00, 2020-01-31 23:00:00, 2020-01-31 23:00:00, 2020-01-31 23:…
#> $ time_period_time_interval_end   <dttm> 2020-02-29 23:00:00, 2020-02-29 23:00:00, 2020-02-29 23:00:00, 2020-02-29 23:…
#> $ ts_resolution                   <chr> "PT60M", "PT60M", "PT60M", "PT60M", "PT60M", "PT60M", "PT60M", "PT60M", "PT60M…
#> $ ts_time_interval_start          <dttm> 2020-01-31 23:00:00, 2020-01-31 23:00:00, 2020-01-31 23:00:00, 2020-01-31 23:…
#> $ ts_time_interval_end            <dttm> 2020-02-29 23:00:00, 2020-02-29 23:00:00, 2020-02-29 23:00:00, 2020-02-29 23:…
#> $ ts_mrid                         <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ ts_point_dt_start               <dttm> 2020-01-31 23:00:00, 2020-02-01 00:00:00, 2020-02-01 01:00:00, 2020-02-01 02:…
#> $ ts_point_quantity               <dbl> 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 7.20, 431.69, 1244.42, 1843.30…
#> $ ts_quantity_measure_unit_name   <chr> "MAW", "MAW", "MAW", "MAW", "MAW", "MAW", "MAW", "MAW", "MAW", "MAW", "MAW", "…
dplyr::glimpse(df_list$`Intraday`)
#> Rows: 986
#> Columns: 23
#> $ ts_in_bidding_zone_domain_mrid  <chr> "10YFR-RTE------C", "10YFR-RTE------C", "10YFR-RTE------C", "10YFR-RTE------C"…
#> $ ts_in_bidding_zone_domain_name  <chr> "France", "France", "France", "France", "France", "France", "France", "France"…
#> $ type                            <chr> "A69", "A69", "A69", "A69", "A69", "A69", "A69", "A69", "A69", "A69", "A69", "…
#> $ type_def                        <chr> "Wind and solar forecast", "Wind and solar forecast", "Wind and solar forecast…
#> $ process_type                    <chr> "A40", "A40", "A40", "A40", "A40", "A40", "A40", "A40", "A40", "A40", "A40", "…
#> $ process_type_def                <chr> "Intraday process", "Intraday process", "Intraday process", "Intraday process"…
#> $ ts_object_aggregation           <chr> "A08", "A08", "A08", "A08", "A08", "A08", "A08", "A08", "A08", "A08", "A08", "…
#> $ ts_object_aggregation_def       <chr> "Resource type", "Resource type", "Resource type", "Resource type", "Resource …
#> $ ts_business_type                <chr> "A94", "A94", "A94", "A94", "A94", "A94", "A94", "A94", "A94", "A94", "A94", "…
#> $ ts_business_type_def            <chr> "Solar generation", "Solar generation", "Solar generation", "Solar generation"…
#> $ ts_mkt_psr_type                 <chr> "B16", "B16", "B16", "B16", "B16", "B16", "B16", "B16", "B16", "B16", "B16", "…
#> $ ts_mkt_psr_type_def             <chr> "Solar unspecified", "Solar unspecified", "Solar unspecified", "Solar unspecif…
#> $ created_date_time               <dttm> 2026-03-09 12:42:15, 2026-03-09 12:42:15, 2026-03-09 12:42:15, 2026-03-09 12:…
#> $ revision_number                 <dbl> 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, …
#> $ time_period_time_interval_start <dttm> 2020-02-01 06:00:00, 2020-02-01 06:00:00, 2020-02-01 06:00:00, 2020-02-01 06:…
#> $ time_period_time_interval_end   <dttm> 2020-02-29 23:00:00, 2020-02-29 23:00:00, 2020-02-29 23:00:00, 2020-02-29 23:…
#> $ ts_resolution                   <chr> "PT60M", "PT60M", "PT60M", "PT60M", "PT60M", "PT60M", "PT60M", "PT60M", "PT60M…
#> $ ts_time_interval_start          <dttm> 2020-02-01 06:00:00, 2020-02-01 06:00:00, 2020-02-01 06:00:00, 2020-02-01 06:…
#> $ ts_time_interval_end            <dttm> 2020-02-01 23:00:00, 2020-02-01 23:00:00, 2020-02-01 23:00:00, 2020-02-01 23:…
#> $ ts_mrid                         <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ ts_point_dt_start               <dttm> 2020-02-01 06:00:00, 2020-02-01 07:00:00, 2020-02-01 08:00:00, 2020-02-01 09:…
#> $ ts_point_quantity               <dbl> 0.00, 6.74, 414.99, 1145.83, 1751.88, 2224.50, 2442.92, 2361.13, 1889.89, 1107…
#> $ ts_quantity_measure_unit_name   <chr> "MAW", "MAW", "MAW", "MAW", "MAW", "MAW", "MAW", "MAW", "MAW", "MAW", "MAW", "…
dplyr::glimpse(df_list$`Current`)
#> Rows: 1,392
#> Columns: 23
#> $ ts_in_bidding_zone_domain_mrid  <chr> "10YFR-RTE------C", "10YFR-RTE------C", "10YFR-RTE------C", "10YFR-RTE------C"…
#> $ ts_in_bidding_zone_domain_name  <chr> "France", "France", "France", "France", "France", "France", "France", "France"…
#> $ type                            <chr> "A69", "A69", "A69", "A69", "A69", "A69", "A69", "A69", "A69", "A69", "A69", "…
#> $ type_def                        <chr> "Wind and solar forecast", "Wind and solar forecast", "Wind and solar forecast…
#> $ process_type                    <chr> "A18", "A18", "A18", "A18", "A18", "A18", "A18", "A18", "A18", "A18", "A18", "…
#> $ process_type_def                <chr> "Intraday total", "Intraday total", "Intraday total", "Intraday total", "Intra…
#> $ ts_object_aggregation           <chr> "A08", "A08", "A08", "A08", "A08", "A08", "A08", "A08", "A08", "A08", "A08", "…
#> $ ts_object_aggregation_def       <chr> "Resource type", "Resource type", "Resource type", "Resource type", "Resource …
#> $ ts_business_type                <chr> "A94", "A94", "A94", "A94", "A94", "A94", "A94", "A94", "A94", "A94", "A94", "…
#> $ ts_business_type_def            <chr> "Solar generation", "Solar generation", "Solar generation", "Solar generation"…
#> $ ts_mkt_psr_type                 <chr> "B16", "B16", "B16", "B16", "B16", "B16", "B16", "B16", "B16", "B16", "B16", "…
#> $ ts_mkt_psr_type_def             <chr> "Solar unspecified", "Solar unspecified", "Solar unspecified", "Solar unspecif…
#> $ created_date_time               <dttm> 2026-03-09 12:42:16, 2026-03-09 12:42:16, 2026-03-09 12:42:16, 2026-03-09 12:…
#> $ revision_number                 <dbl> 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, …
#> $ time_period_time_interval_start <dttm> 2020-01-31 23:00:00, 2020-01-31 23:00:00, 2020-01-31 23:00:00, 2020-01-31 23:…
#> $ time_period_time_interval_end   <dttm> 2020-02-29 23:00:00, 2020-02-29 23:00:00, 2020-02-29 23:00:00, 2020-02-29 23:…
#> $ ts_resolution                   <chr> "PT60M", "PT60M", "PT60M", "PT60M", "PT60M", "PT60M", "PT60M", "PT60M", "PT60M…
#> $ ts_time_interval_start          <dttm> 2020-01-31 23:00:00, 2020-01-31 23:00:00, 2020-01-31 23:00:00, 2020-01-31 23:…
#> $ ts_time_interval_end            <dttm> 2020-02-29 23:00:00, 2020-02-29 23:00:00, 2020-02-29 23:00:00, 2020-02-29 23:…
#> $ ts_mrid                         <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ ts_point_dt_start               <dttm> 2020-01-31 23:00:00, 2020-02-01 00:00:00, 2020-02-01 01:00:00, 2020-02-01 02:…
#> $ ts_point_quantity               <dbl> 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 6.74, 414.99, 1512.90, 2320.00…
#> $ ts_quantity_measure_unit_name   <chr> "MAW", "MAW", "MAW", "MAW", "MAW", "MAW", "MAW", "MAW", "MAW", "MAW", "MAW", "…