Get Netted Volumes (IFs IN 3.10)

The net position informs whether the given area imports or exports energy. Those rows which hold the queried eic value in the 'ts_connecting_domain_mrid' column show the export value. Those rows which hold the queried eic value in the 'ts_acquiring_domain_mrid' column show the import value.

Usage

netted_volumes(
  eic = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 7L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)

Arguments

eic

Energy Identification Code of the area

period_start

POSIXct or YYYY-MM-DD HH:MM:SS format One day range limit applies

period_end

POSIXct or YYYY-MM-DD HH:MM:SS format One day range limit applies

tidy_output

Defaults to TRUE. flatten nested tables

security_token

Security token for ENTSO-E transparency platform

Value

A tibble::tibble() with the queried data, or NULL if no data is available for the given parameters.

Examples

df <- entsoeapi::netted_volumes(
  eic = "10YCZ-CEPS-----N",
  period_start = lubridate::ymd(x = "2022-08-16", tz = "CET"),
  period_end = lubridate::ymd(x = "2022-08-17", tz = "CET"),
  tidy_output = TRUE
)
#> 
#> ── API call ────────────────────────────────────────────────────────────────────────────────────────────────────────────
#> → https://web-api.tp.entsoe.eu/api?documentType=B17&processType=A63&Acquiring_domain=10YCZ-CEPS-----N&Connecting_Domain=10YCZ-CEPS-----N&periodStart=202208152200&periodEnd=202208162200&securityToken=<...>
#> <- HTTP/2 200 
#> <- date: Tue, 17 Mar 2026 22:11:26 GMT
#> <- content-type: application/zip
#> <- content-length: 1784
#> <- content-disposition: attachment; filename="Netted and Exchanged Volumes [IFs IN 3.10, mFRR 3.17, aFRR 3.16]_202208152200-202208162200.zip"
#> <- x-content-type-options: nosniff
#> <- x-xss-protection: 0
#> <- strict-transport-security: max-age=15724800; includeSubDomains
#> <- 
#> ✔ response has arrived
#> ✔ /tmp/RtmpOgumiW/001-NETTED_AND_EXCHANGED_VOLUMES_202208152200-202208162200.xml has been read in
#> ℹ pulling area_eic_name table from cache

dplyr::glimpse(df)
#> Rows: 96
#> Columns: 23
#> $ area_domain_mrid              <chr> "10Y1001C--00119X", "10Y1001C--00119X", "10Y1001C--00119X", "10Y1001C--00119X", …
#> $ area_domain_name              <chr> "Imbalance netting region", "Imbalance netting region", "Imbalance netting regio…
#> $ ts_connecting_domain_mrid     <chr> "10Y1001C--00119X", "10Y1001C--00119X", "10Y1001C--00119X", "10Y1001C--00119X", …
#> $ ts_connecting_domain_name     <chr> "Imbalance netting region", "Imbalance netting region", "Imbalance netting regio…
#> $ ts_acquiring_domain_mrid      <chr> "10YCZ-CEPS-----N", "10YCZ-CEPS-----N", "10YCZ-CEPS-----N", "10YCZ-CEPS-----N", …
#> $ ts_acquiring_domain_name      <chr> "Czech Republic", "Czech Republic", "Czech Republic", "Czech Republic", "Czech R…
#> $ doc_status_value              <chr> "A02", "A02", "A02", "A02", "A02", "A02", "A02", "A02", "A02", "A02", "A02", "A0…
#> $ doc_status                    <chr> "Allocated capacity schedule", "Allocated capacity schedule", "Allocated capacit…
#> $ type                          <chr> "B17", "B17", "B17", "B17", "B17", "B17", "B17", "B17", "B17", "B17", "B17", "B1…
#> $ type_def                      <chr> "Aggregated netted external TSO schedule document", "Aggregated netted external …
#> $ process_type                  <chr> "A63", "A63", "A63", "A63", "A63", "A63", "A63", "A63", "A63", "A63", "A63", "A6…
#> $ process_type_def              <chr> "Imbalance Netting", "Imbalance Netting", "Imbalance Netting", "Imbalance Nettin…
#> $ ts_business_type              <chr> "B09", "B09", "B09", "B09", "B09", "B09", "B09", "B09", "B09", "B09", "B09", "B0…
#> $ ts_business_type_def          <chr> "Net position", "Net position", "Net position", "Net position", "Net position", …
#> $ created_date_time             <dttm> 2026-03-17 22:11:26, 2026-03-17 22:11:26, 2026-03-17 22:11:26, 2026-03-17 22:11…
#> $ revision_number               <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,…
#> $ ts_resolution                 <chr> "PT15M", "PT15M", "PT15M", "PT15M", "PT15M", "PT15M", "PT15M", "PT15M", "PT15M",…
#> $ ts_time_interval_start        <dttm> 2022-08-15 22:00:00, 2022-08-15 22:00:00, 2022-08-15 22:00:00, 2022-08-15 22:00…
#> $ ts_time_interval_end          <dttm> 2022-08-15 23:45:00, 2022-08-15 23:45:00, 2022-08-15 23:45:00, 2022-08-15 23:45…
#> $ ts_mrid                       <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,…
#> $ ts_point_dt_start             <dttm> 2022-08-15 22:00:00, 2022-08-15 22:15:00, 2022-08-15 22:30:00, 2022-08-15 22:45…
#> $ ts_point_quantity             <dbl> 1.1357, 3.7387, 9.6364, 13.7141, 5.7463, 3.1041, 5.5129, 1.2547, 2.4147, 6.8177,…
#> $ ts_quantity_measure_unit_name <chr> "MWH", "MWH", "MWH", "MWH", "MWH", "MWH", "MWH", "MWH", "MWH", "MWH", "MWH", "MW…